change exchangerate type to decimal (18,4)

+++ SQL_DEV/silver_mm/silver_mm_initial/silver_zmmpurchasingdoc_initial.sql
@@ -22,7 +22,7 @@ SELECT DISTINCT
     CAST(R.purchasinggroup AS STRING) AS purchasinggroup,
     CAST(G.purchasinggroupname as STRING) AS purchasinggroupname,
     CAST(R.documentcurrency AS STRING) AS documentcurrency,
-    CAST(R.exchangerate AS double) AS exchangerate,
+    CAST(R.exchangerate AS DECIMAL(18,4)) AS exchangerate,
     to_date(R.purchasingdocumentorderdate, 'yyyyMMdd') AS purchasingdocumentorderdate,
     CAST(R.paymentterms AS STRING) AS paymentterms,
     CAST(R.incotermsclassification AS STRING) AS incotermsclassification,
